Split XTypes test into Pub and Sub projects
