+++ tests/DCPS/XTypes/.gitignore
@@ -1,14 +1,29 @@
-/XTypes
-/XTypesC.cpp
-/XTypesC.h
-/XTypesC.inl
-/XTypesS.cpp
-/XTypesS.h
-/XTypesTypeSupport.idl
-/XTypesTypeSupportC.cpp
-/XTypesTypeSupportC.h
-/XTypesTypeSupportC.inl
-/XTypesTypeSupportImpl.cpp
-/XTypesTypeSupportImpl.h
-/XTypesTypeSupportS.cpp
-/XTypesTypeSupportS.h
+/XTypesPub
+/XTypesPubC.cpp
+/XTypesPubC.h
+/XTypesPubC.inl
+/XTypesPubS.cpp
+/XTypesPubS.h
+/XTypesPubTypeSupport.idl
+/XTypesPubTypeSupportC.cpp
+/XTypesPubTypeSupportC.h
+/XTypesPubTypeSupportC.inl
+/XTypesPubTypeSupportImpl.cpp
+/XTypesPubTypeSupportImpl.h
+/XTypesPubTypeSupportS.cpp
+/XTypesPubTypeSupportS.h
+
+/XTypesSub
+/XTypesSubC.cpp
+/XTypesSubC.h
+/XTypesSubC.inl
+/XTypesSubS.cpp
+/XTypesSubS.h
+/XTypesSubTypeSupport.idl
+/XTypesSubTypeSupportC.cpp
+/XTypesSubTypeSupportC.h
+/XTypesSubTypeSupportC.inl
+/XTypesSubTypeSupportImpl.cpp
+/XTypesSubTypeSupportImpl.h
+/XTypesSubTypeSupportS.cpp
+/XTypesSubTypeSupportS.h
