+++ tests/DCPS/XTypes/XTypesSub.idl
@@ -1,0 +1,85 @@
+#include <tao/OctetSeq.pidl>
+
+@topic
+struct PlainCdrStruct {
+  @key long key;
+  long value;
+};
+
+// Final structues
+
+@topic
+@final
+struct FinalStructSub {
+  @key long key;
+};
+
+// Appendable structues
+
+@topic
+@appendable
+struct AppendableStruct {
+  @key long key;
+};
+
+// Mutable structues
+
+@topic
+@mutable
+struct MutableStruct {
+  @id(1) @key long key;
+  @id(2) long additional_field;
+};
+
+
+// Unions
+
+enum UnionDisc {
+  E_KEY,
+  E_OCTET_FIELD,
+  E_ADDITIONAL_FIELD
+};
+
+enum ModifiedUnionDisc {
+  E_KEY_MODIFIED,
+  E_OCTET_FIELD_MODIFIED,
+  E_ADDITIONAL_FIELD_MODIFIED
+};
+
+@topic
+@mutable
+union MutableUnion switch (UnionDisc) {
+case E_KEY:
+  @id(6) long key;
+case E_OCTET_FIELD:
+  @id(8) octet octet_field;
+case E_ADDITIONAL_FIELD:
+  @id(12) long additional_field;
+};
+
+// Tryconstruct
+
+typedef string<20> str20;
+typedef string<64> str64;
+
+@topic
+@mutable
+struct Trim20Struct {
+  @try_construct(TRIM) str20 trim_string;
+};
+
+// Dependencies
+
+@appendable
+struct NestedStruct {
+  long additional_field;
+};
+
+// No XTypes Test
+
+@topic
+@appendable
+struct AppendableStructNoXTypes {
+  @key long key;
+};
+
